add semantic elements, improve accessibility, more consistent styling
