+++ frontend/src/App.tsx
@@ -1,6 +1,4 @@
-import NavBar from "./components/NavBar.jsx";
-
-import "./App.css";
+import NavBar from "./components/NavBar.js";
 
 function App() {
   return (
