+++ frontend/src/components/Button.tsx
@@ -1,5 +1,5 @@
 import styles from "./Button.module.css";
-export default function Button(props) {
+export default function Button(props: { onClick: any; children: String }) {
   const { onClick, children } = props;
   return (
     <button className={styles.button} onClick={onClick}>
